Some attempts to cast composite types to binary
This is problematic because postgres is quite strict on the field types
and doesn't apply normal cast rules, e.g. between integer types. So we
need to configure every field to a dumper, but knowing the target oids
rather than the source classes, which get in the way with what we
have so far.

Putting this on hold for the moment and focusing on something else.